requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ./requirements/base.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
billiard==4.2.1
    # via celery
celery==5.3.1
    # via -r ./requirements/base.in
certifi==2024.8.30
    # via sentry-sdk
cffi==1.17.1
    # via
    #   cryptography
    #   gevent
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   redis-om
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cryptography==43.0.1
    # via
    #   types-pyopenssl
    #   types-redis
django==4.2.3
    # via
    #   -r ./requirements/base.in
    #   django-cachalot
    #   django-redis
    #   djangorestframework
django-cachalot==2.6.1
    # via -r ./requirements/base.in
django-redis==5.3.0
    # via -r ./requirements/base.in
djangorestframework==3.14.0
    # via -r ./requirements/base.in
gevent==24.2.1
    # via -r ./requirements/base.in
greenlet==3.1.1
    # via gevent
gunicorn==21.2.0
    # via -r ./requirements/base.in
hiredis==2.4.0
    # via redis-om
kombu==5.4.2
    # via celery
more-itertools==9.1.0
    # via redis-om
packaging==24.1
    # via gunicorn
prompt-toolkit==3.0.48
    # via click-repl
psycopg2==2.9.9
    # via -r ./requirements/base.in
pycparser==2.22
    # via cffi
pydantic==2.0.3
    # via redis-om
pydantic-core==2.3.0
    # via pydantic
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r ./requirements/base.in
python-ulid==1.1.0
    # via redis-om
pytz==2024.2
    # via djangorestframework
redis==4.6.0
    # via
    #   -r ./requirements/base.in
    #   django-redis
    #   redis-om
redis-om==0.2.1
    # via -r ./requirements/base.in
sentry-sdk==1.28.1
    # via -r ./requirements/base.in
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
types-cffi==1.16.0.20240331
    # via types-pyopenssl
types-pyopenssl==24.1.0.20240722
    # via types-redis
types-redis==4.6.0.20241004
    # via redis-om
types-setuptools==75.1.0.20240917
    # via types-cffi
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   redis-om
tzdata==2024.2
    # via
    #   celery
    #   django
    #   kombu
urllib3==2.2.3
    # via sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
zope-event==5.0
    # via gevent
zope-interface==7.0.3
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
